refactor(sanity): update imports
