+++ packages/sanity/src/_singletons/core/form/studio/TreeEditingEnabledContext.ts
@@ -1,6 +1,6 @@
 import {createContext} from 'react'
 
-import type {TreeEditingEnabledContextValue} from './types'
+import type {TreeEditingEnabledContextValue} from '../../../../core/form/studio/tree-editing/context/enabled/useTreeEditingEnabled'
 
 /**
  * @internal
