--- packages/sanity/src/_singletons/core/form/tree-editing/index.ts
@@ -1,2 +1,0 @@
-export * from './TreeEditingEnabledContext'
-export * from './types'
