--- packages/sanity/src/_singletons/core/form/tree-editing/types.ts
@@ -1,13 +1,0 @@
-/**
- * @internal
- */
-export interface TreeEditingEnabledContextValue {
-  /**
-   * A boolean indicating whether tree editing is enabled
-   */
-  enabled: boolean
-  /**
-   * A boolean indicating whether legacy editing is enabled
-   */
-  legacyEditing: boolean
-}
