+++ packages/sanity/src/_singletons/index.ts
@@ -21,8 +21,8 @@ export * from './core/form/studio/FormCallbacksContext'
 export * from './core/form/studio/PresenceContext'
 export * from './core/form/studio/ReferenceInputOptionsContext'
 export * from './core/form/studio/reviewChanges/ReviewChangesContext'
+export * from './core/form/studio/TreeEditingEnabledContext'
 export * from './core/form/studio/ValidationContext'
-export * from './core/form/tree-editing'
 export * from './core/i18n/LocaleContext'
 export * from './core/presence/FormFieldPresenceContext'
 export * from './core/schedulePublishing/DocumentActionPropsContext'
