+++ packages/sanity/src/core/form/studio/tree-editing/context/enabled/useTreeEditingEnabled.ts
@@ -1,5 +1,19 @@
 import {useContext} from 'react'
-import {TreeEditingEnabledContext, type TreeEditingEnabledContextValue} from 'sanity/_singletons'
+import {TreeEditingEnabledContext} from 'sanity/_singletons'
+
+/**
+ * @internal
+ */
+export interface TreeEditingEnabledContextValue {
+  /**
+   * A boolean indicating whether tree editing is enabled
+   */
+  enabled: boolean
+  /**
+   * A boolean indicating whether legacy editing is enabled
+   */
+  legacyEditing: boolean
+}
 
 /**
  * @internal
